clean up cruft
git-svn-id: https://svn.codehaus.org/mule/trunk/mule@770 bf997673-6b11-0410-b953-e057580c5b09

+++ src/java/org/mule/util/SgmlCodec.java
@@ -295,7 +295,7 @@ public class SgmlCodec {
         for (int i = 0; i < length; i++) {
 
             char c = s.charAt(i);
-            int j = (int) c;
+            int j = c;
 
             if ((j < 0x100) && (encoder[j] != null)) {
